// src/server/database/interfaces/reports.interface.ts
export interface IReportsWithAccountedFinances {
  readonly userId: string;
  readonly dateFrom: string;
  readonly dateTo: string;
  readonly reportId: number;
  readonly tax: number;
  readonly financesAccountedAt: number;
  readonly profit: number;
  readonly margin: number;
  readonly productCosts: number;
  readonly insuranceFee: number;
  readonly additionalInsuranceFee: number;
}

export interface IReports extends Document {
  readonly userId: string;
  readonly reports: Report[];
  readonly schemaVersion?: number;
  readonly reportsWithAccountedFinances: IReportsWithAccountedFinances[];
}